Move AnyExplicitScript to primitive
