+++ lib/primitive/cardano-wallet-primitive.cabal
@@ -112,6 +112,7 @@ library
     Cardano.Wallet.Primitive.SyncProgress
     Cardano.Wallet.Primitive.Types.Address
     Cardano.Wallet.Primitive.Types.Address.Gen
+    Cardano.Wallet.Primitive.Types.AnyExplicitScripts
     Cardano.Wallet.Primitive.Types.Block
     Cardano.Wallet.Primitive.Types.Block.Gen
     Cardano.Wallet.Primitive.Types.BlockSummary
